+++ example/android/app/src/main/kotlin/com/example/custom_error_demo/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.custom_error_demo
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}